fix: move version provider logic in separate script from handler
