+++ packages/aws-rfdk/lib/core/lambdas/nodejs/version-provider/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from './handler';
+export * from './version-provider';
